@@ -21,28 +21,29 @@
 		"url": "git+https://github.com/religiosa1/notifier.git"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^2.1.0",
-		"@sveltejs/adapter-node": "^1.3.1",
-		"@sveltejs/kit": "^1.24.1",
-		"@typescript-eslint/eslint-plugin": "^6.6.0",
-		"@typescript-eslint/parser": "^6.6.0",
+		"@sveltejs/adapter-auto": "^3.0.0",
+		"@sveltejs/adapter-node": "^2.0.0",
+		"@sveltejs/kit": "^2.0.0",
+		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@typescript-eslint/eslint-plugin": "^6.14.0",
+		"@typescript-eslint/parser": "^6.14.0",
 		"dotenv": "^16.3.1",
-		"eslint": "^8.49.0",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-svelte": "^2.33.1",
-		"prettier": "^3.0.3",
-		"prettier-plugin-svelte": "^3.0.3",
-		"sass": "^1.66.1",
-		"svelte": "^4.2.0",
-		"svelte-check": "^3.5.1",
+		"eslint": "^8.55.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.35.1",
+		"prettier": "^3.1.1",
+		"prettier-plugin-svelte": "^3.1.2",
+		"sass": "^1.69.5",
+		"svelte": "^4.2.8",
+		"svelte-check": "^3.6.2",
 		"tslib": "^2.6.2",
-		"typescript": "^5.2.2",
-		"vite": "^4.4.9",
-		"vitest": "^0.34.4"
+		"typescript": "^5.3.3",
+		"vite": "^5.0.0",
+		"vitest": "^1.0.0"
 	},
 	"type": "module",
 	"dependencies": {
-		"nanoid": "^4.0.2",
-		"zod": "^3.22.2"
+		"nanoid": "^5.0.4",
+		"zod": "^3.22.4"
 	}
 }
