@@ -23,38 +23,41 @@
     "url": "git+https://github.com/religiosa1/notifier.git"
   },
   "dependencies": {
-    "@fastify/basic-auth": "^5.0.0",
-    "@fastify/cookie": "^9.0.4",
-    "@fastify/jwt": "^7.2.1",
+    "@fastify/basic-auth": "^5.1.1",
+    "@fastify/cookie": "^9.2.0",
+    "@fastify/jwt": "^7.2.4",
     "@fastify/middie": "^8.3.0",
-    "@tsconfig/node-lts": "^18.12.5",
+    "@hono/zod-validator": "^0.1.11",
+    "@tsconfig/node-lts": "^20.1.0",
     "@tsconfig/strictest": "^2.0.2",
-    "@types/connect": "^3.4.36",
+    "@types/connect": "^3.4.38",
     "bcrypt": "^5.1.1",
     "dotenv": "^16.3.1",
-    "drizzle-orm": "^0.28.6",
-    "fastify": "^4.22.2",
+    "drizzle-orm": "^0.29.1",
+    "fastify": "^4.25.1",
     "fastify-plugin": "^4.5.1",
     "fastify-type-provider-zod": "^1.1.9",
     "jsonc-parser": "^3.2.0",
-    "node-telegram-bot-api": "^0.61.0",
-    "pino": "^8.15.1",
-    "postgres": "^3.3.5",
-    "rfc4648": "^1.5.2",
-    "zod": "^3.22.2"
+    "node-telegram-bot-api": "^0.64.0",
+    "pino": "^8.17.1",
+    "postgres": "^3.4.3",
+    "rfc4648": "^1.5.3",
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/bcrypt": "^5.0.0",
-    "@types/jest": "^29.5.4",
-    "@types/node": "^20.6.0",
-    "@types/node-telegram-bot-api": "^0.61.7",
+    "@hono/node-server": "^1.3.3",
+    "@types/bcrypt": "^5.0.2",
+    "@types/jest": "^29.5.11",
+    "@types/node": "^20.10.5",
+    "@types/node-telegram-bot-api": "^0.64.1",
     "cross-env": "^7.0.3",
-    "drizzle-kit": "^0.19.13",
-    "jest": "^29.6.4",
+    "drizzle-kit": "^0.20.6",
+    "hono": "^3.11.8",
+    "jest": "^29.7.0",
     "ts-jest": "^29.1.1",
     "tsconfig-paths": "^4.2.0",
-    "tsup": "^7.2.0",
-    "tsx": "^3.12.8",
-    "typescript": "^5.2.2"
+    "tsup": "^8.0.1",
+    "tsx": "^4.6.2",
+    "typescript": "^5.3.3"
   }
 }
