@@ -15,12 +15,12 @@
     "url": "git+https://github.com/religiosa1/notifier.git"
   },
   "dependencies": {
-    "zod": "^3.22.2"
+    "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/jest": "^29.5.4",
-    "jest": "^29.6.4",
+    "@types/jest": "^29.5.11",
+    "jest": "^29.7.0",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.3.3"
   }
 }
