@@ -1,12 +1,13 @@
 import { esc } from "src/util/esc";
 import type { Bot } from "./Bot";
 import type { IBot } from "./Models";
-import { BaseLogger } from "pino";
-import { inject } from "src/injection";
+import type { BaseLogger } from "pino";
+import { di } from "src/injection";
+
 
 /** Basic mock of Bot class for testing purposes */
 export class BotMock implements IBot {
-	constructor(token: string, private readonly logger: BaseLogger = inject("logger")) {
+	constructor(token: string, private readonly logger: BaseLogger = di.inject("logger")) {
 		if (!token) {
 			throw new Error("Bot token isn't present in the env variables");
 		}
