@@ -10,17 +10,19 @@ export const handleFetch: HandleFetch = ({ event, request, fetch }) => {
 		request.headers.set('Authorization', event.cookies.get('Authorization') || "");
 		request.headers.set('Content-Type', 'application/json');
 	}
-
-	if (event.url.pathname === "/login") {
-		return fetch(request);
-	}
 	return fetch(request).then(r => {
 		// We don't intercept settings requests with redirects
 		if (request.url.includes(serverUrl("/settings").toString())) {
 			return r;
 		}
-		if (r.status === 403 || r.status === 401) {
-			throw redirect(303, base + uri`/login?referer=${event.url.pathname}`);
+		// Special status code for uninitialized server
+		if (r.status === 550) {
+			// TODO reset global_hasValidServerSettings if this happened
+			redirect(303, base + uri`/setup`);
+		}
+		if (event.url.pathname !== "/login" && r.status === 403 || r.status === 401) {
+			event.cookies.delete("Authorization", { path: "/"});
+			redirect(303, base + uri`/login?referer=${event.url.pathname}`);
 		}
 		return r;
 	});
