@@ -5,6 +5,7 @@ import { base } from "$app/paths";
 import { uri } from "~/helpers/uri";
 import { serverUrl } from "~/helpers/serverUrl";
 
+// TODO also drop the status in fetch hooks, move it to service
 let global_hasValidServerSettings = false;
 
 export const load: LayoutServerLoad = async ({ cookies, url, fetch }) => {
@@ -14,11 +15,13 @@ export const load: LayoutServerLoad = async ({ cookies, url, fetch }) => {
 		if (url.pathname === "/setup") {
 			return;
 		}
-		global_hasValidServerSettings = await fetch(serverUrl('/settings')).then(r => {
-			return r.ok || r.status === 401 // 401 is possible only if settings were created
-		}, () => false);
-		if (!global_hasValidServerSettings) {
-			throw redirect(303, base + `/setup`);
+		const settingsResponse = await fetch(serverUrl('/settings'));
+		if (settingsResponse.status === 550) {
+			redirect(303, base + `/setup`);
+		} else if (settingsResponse.ok || settingsResponse.status === 401) {
+			// 401 means that the user wasn't unauthorized, but this error can only be thrown
+			// if server has been initialized with a config
+			global_hasValidServerSettings = true;
 		}
 	}
 
@@ -26,11 +29,14 @@ export const load: LayoutServerLoad = async ({ cookies, url, fetch }) => {
 		if (url.pathname === "/login") {
 			return {};
 		} else {
-			throw redirect(303, base + uri`/login?referer=${url.pathname}`);
+			redirect(303, base + uri`/login?referer=${url.pathname}`);
 		}
 	}
-	const tokenPayload = decodeJWT(auth);
-	return {
-		user: tokenPayload,
-	};
+	try {
+		const tokenPayload = decodeJWT(auth);
+		return { user: tokenPayload };
+	} catch {
+		cookies.delete("Authorization", { path: "/" });
+		redirect(303, base + uri`/login?referer=${url.pathname}`);
+	}
 }
