@@ -9,7 +9,7 @@ export const load: PageServerLoad = async ({ parent }) => {
   const { user } = await parent();
 
   if (user) {
-    throw redirect(303, base + '/');
+    redirect(303, base + '/');
   }
 }
 
@@ -28,11 +28,10 @@ export const actions: Actions = {
       console.error("Login error", e);
       return handleActionFailure(e, { name });
     }
-
     if (!hasField(serverData, "token", "string")) {
-      throw fail(500, { error: "Bad server response, missing `token` field." });
+      return fail(500, { error: "Bad server response, missing `token` field." });
     }
-    cookies.set("Authorization", `Bearer ${serverData.token}`);
-    throw redirect(303, url.searchParams.get("referer") || "/");
+    cookies.set("Authorization", `Bearer ${serverData.token}`, { path: '/' });
+    redirect(303, url.searchParams.get("referer") || "/");
   }
 };
