@@ -1,7 +1,7 @@
+import pino, { type BaseLogger } from "pino";
 import { AppListenService } from "src/services/AppListenService";
+import { BotService } from "./services/BotService";
 import { SettingsService } from "src/services/SettingsService";
-import pino, { BaseLogger } from "pino";
-import { IBot } from "src/Bot/Models";
 import { DatabaseConnectionManager } from "src/db/DatabaseConnectionManager";
 import { GroupsRepository } from "src/repositories/GroupsRepository";
 import { UserToGroupRelationsRepository } from "src/repositories/UserToGroupRelationsRepository";
@@ -11,48 +11,42 @@ import { ChannelsRepository } from "src/repositories/ChannelsRepository";
 import { ChannelToGroupRelationsRepository } from "src/repositories/ChannelToGroupRelationsRepository";
 import { UserConfirmationRequestsRepository } from "src/repositories/UserConfirmationRequestsRepository";
 import { UserToChannelRelationsRepository } from "src/repositories/UserToChannelRelationsRepository";
+import { DatabaseMigrator } from "./db/DatabaseMigrator";
 
-export const container = {
-	SettingsService: undefined as unknown as SettingsService,
-	AppListenService: new AppListenService(),
-	logger: pino({ level: 'info' }) as BaseLogger,
-	Bot: undefined as IBot | undefined,
-	db: undefined as unknown as DatabaseConnectionManager,
-	GroupsRepository: undefined as unknown as GroupsRepository,
-	UsersRepository: undefined as unknown as UsersRepository,
-	UserConfirmationRequestsRepository: undefined as unknown as UserConfirmationRequestsRepository,
-	UserToGroupRelationsRepository: undefined as unknown as UserToGroupRelationsRepository,
-	ApiKeysRepository: undefined as unknown as ApiKeysRepository,
-	ChannelsRepository: undefined as unknown as ChannelsRepository,
-	ChannelToGroupRelationsRepository: undefined as unknown as ChannelToGroupRelationsRepository,
-	UserToChannelRelationsRepository: undefined as unknown as UserToChannelRelationsRepository
-};
-type Container = typeof container;
-
-container.SettingsService = new SettingsService();
-container.db = new DatabaseConnectionManager();
-container.GroupsRepository = new GroupsRepository();
-container.UsersRepository = new UsersRepository();
-container.UserConfirmationRequestsRepository = new UserConfirmationRequestsRepository();
-container.UserToGroupRelationsRepository = new UserToGroupRelationsRepository();
-container.ApiKeysRepository = new ApiKeysRepository();
-container.ChannelsRepository = new ChannelsRepository();
-container.ChannelToGroupRelationsRepository = new ChannelToGroupRelationsRepository();
-container.UserToChannelRelationsRepository = new UserToChannelRelationsRepository();
-
-export function inject<TItem extends keyof Container>(key: TItem): Container[TItem] {
-	if (!(key in container)) {
-		throw new Error("Incorrect injection key");
-	}
-	return container[key];
+type ContainerInit<T> = {
+	[K in keyof T]: () => T[K];
 }
 
-export function register<TItem extends keyof Container>(
-	key: TItem,
-	item: Container[TItem],
-): void {
-	if (!(key in container)) {
-		throw new Error("Incorrect injection key");
+class DiContainer<T extends Record<string, {}>> {
+	private readonly container: Partial<T> = {};
+	constructor(private readonly containerInit: ContainerInit<T>) {
+    this.inject = this.inject.bind(this);
+		
 	}
-	container[key] = item;
+
+	inject<TItem extends keyof T & string>(key: TItem): T[TItem] {
+		if (!(key in this.containerInit)) {
+			throw new Error(`Incorrect injection key during inject: '${key}'`);
+		}
+		this.container[key] ??= this.containerInit[key]();
+		return this.container[key] as T[TItem] ;
+	}
 }
+
+export const di = new DiContainer({
+	AppListenService: () => new AppListenService(),
+	logger: () => pino({ level: 'info' }) as BaseLogger,
+	SettingsService: () => new SettingsService(),
+	db: () => new DatabaseConnectionManager(),
+	DatabaseMigrator: () => new DatabaseMigrator(),
+	GroupsRepository: () => new GroupsRepository(),
+	UsersRepository: () => new UsersRepository(),
+	UserConfirmationRequestsRepository: () => new UserConfirmationRequestsRepository(),
+	UserToGroupRelationsRepository: () => new UserToGroupRelationsRepository(),
+	ApiKeysRepository: () => new ApiKeysRepository(),
+	ChannelsRepository: () => new ChannelsRepository(),
+	ChannelToGroupRelationsRepository: () => new ChannelToGroupRelationsRepository(),
+	UserToChannelRelationsRepository: () => new UserToChannelRelationsRepository(),
+	Bot: () => new BotService(),
+});
+
