@@ -1,16 +1,15 @@
-import { inject } from "src/injection";
-import type { Handler } from '@fastify/middie';
-import { IncomingMessage, ServerResponse } from "http";
-import { ConfigUnavailableError } from "src/error/ConfigUnavailableError";
+import { di } from "src/injection";
 
-export const checkSettings: Handler = (req: IncomingMessage, _res: ServerResponse, next: (err?: unknown) => void) => {
-	if (req.url === "/settings") {
-		return next();
+import { ConfigUnavailableError } from "src/error/ConfigUnavailableError";
+import { createMiddleware } from 'hono/factory';
+
+export const checkSettings = createMiddleware((c, next) => {
+	const serverSettings = di.inject("SettingsService");
+	if (c.req.path !== "/settings/setup") {
+		const config = serverSettings.getConfig();
+		if (!config) {
+			throw new ConfigUnavailableError();
+		}
 	}
-	const serverSettings = inject("SettingsService");
-	const config = serverSettings.getConfig();
-	if (!config) {
-		throw new ConfigUnavailableError();
-	}
-	next();
-};
+	return next();
+});
