@@ -1,10 +1,11 @@
-import { Channel, ChannelSubscription } from "@shared/models/Channel";
+import type { Channel, ChannelSubscription } from "@shared/models/Channel";
 import { and, eq, getTableColumns, inArray, sql } from "drizzle-orm";
 import { schema } from "src/db";
-import { inject } from "src/injection";
+import { di } from "src/injection";
+
 
 export class UserToChannelRelationsRepository {
-	private readonly dbm = inject("db");
+	private readonly dbm = di.inject("db");
 
 	//============================================================================
 	// LIST
