@@ -1,56 +1,47 @@
+import { Hono } from 'hono'
 import z from "zod";
-import fp from "fastify-plugin";
-import type { ZodTypeProvider } from "fastify-type-provider-zod";
+import { zValidator } from '@hono/zod-validator'
+import { validationErrorHook } from 'src/middleware/validationErrorHandlers';
+
 import { channelNameSchema } from "@shared/models/Channel";
-import type { IBot } from "src/Bot/Models";
-import { ResultError, result, resultFailureSchema, resultSuccessSchema } from "@shared/models/Result";
-import { inject } from "src/injection";
+import { ResultError } from "@shared/models/Result";
+import { di } from "src/injection";
 
-interface NotifyOptions {
-	bot: IBot
-}
-export default fp<NotifyOptions>(async function (fastify) {
-	const channelsRepository = inject("ChannelsRepository");
+import { authorizeAnyMethod } from 'src/middleware/authorizeAnyMethod';
 
-	fastify.withTypeProvider<ZodTypeProvider>().route({
-		method: "POST",
-		url: "/notify",
-		schema: {
-			body: z.object({
-				channels: z.union([
-					channelNameSchema,
-					z.array(channelNameSchema)
-				]),
-				message: z.string(),
-			}),
-			response: {
-				200: resultSuccessSchema(z.null()),
-				404: resultFailureSchema,
-				422: resultFailureSchema,
-			}
-		},
-		onRequest: fastify.authorizeAnyMethod,
-		async handler(req, reply) {
-			const bot = inject("Bot");
-			if (!bot) {
-				throw new ResultError(503, "Bot isn't initialized");
-			}
+const controller = new Hono();
 
-			const channels = Array.isArray(req.body.channels)
-				? req.body.channels
-				: [req.body.channels];
+controller.use('*', authorizeAnyMethod);
 
-			const chats = await channelsRepository.getUserChatIdsForChannel(channels);
-
-			if (!chats.length) {
-				throw new ResultError(404, "Can't find anyone to send the data to in the provided channels");
-			}
-
-			await bot.broadcastMessage(
-				chats,
-				{ text: req.body.message }
-			);
-			return reply.send(result(null));
+controller.post("/",
+	zValidator("json", z.object({
+		channels: z.union([
+			channelNameSchema,
+			z.array(channelNameSchema),
+		]),
+		message: z.string(),
+	}), validationErrorHook),
+	async (c) => {
+		const channelsRepository = di.inject("ChannelsRepository");
+		const bot = di.inject("Bot").instance;
+		if (!bot) {
+			throw new ResultError(503, "Bot isn't initialized");
 		}
-	});
-});
+		const body = c.req.valid('json');
+		const channels = Array.isArray(body.channels) ? body.channels : [body.channels];
+
+		const chats = await channelsRepository.getUserChatIdsForChannel(channels);
+
+		if (!chats.length) {
+			throw new ResultError(404, "Can't find anyone to send the data to in the provided channels");
+		}
+
+		await bot.broadcastMessage(
+			chats,
+			{ text: body.message }
+		);
+		return c.json(null);
+	}
+);
+
+export default controller;
