@@ -1,109 +1,79 @@
+import { Hono } from 'hono';
 import z from "zod";
-import type { ZodTypeProvider } from "fastify-type-provider-zod";
-import type { FastifyInstance } from "fastify";
-import * as ChannelModel from "@shared/models/Channel";
-import { result, resultFailureSchema, resultSuccessSchema } from "@shared/models/Result";
-import { batchOperationStatsSchema } from "@shared/models/BatchOperationStats";
+import { zValidator } from '@hono/zod-validator';
+import { paramErrorHook, validationErrorHook } from 'src/middleware/validationErrorHandlers';
+
+import type { BatchOperationStats } from "@shared/models/BatchOperationStats";
+import type { Counted } from "@shared/models/Counted";
+import type { ContextVariables } from 'src/ContextVariables';
+import type * as ChannelModel from '@shared/models/Channel';
 import { batchIdsSchema, parseIds } from "@shared/models/batchIds";
-import { paginationDefaults, paginationSchema } from "@shared/models/Pagination";
-import { counted } from "@shared/models/Counted";
-import { inject } from "src/injection";
+import { pageinationQuerySchema, paginationDefaults } from "@shared/models/Pagination";
+import { di } from "src/injection";
 
-export function userChannels<Instace extends FastifyInstance>(fastify: Instace) {
-	const userToChannelRelationsRepository = inject("UserToChannelRelationsRepository");
+import { userIdParamsSchema } from './models';
 
-	const baseUserChannelsUrl = "/users/:userId/channels";
-	const baseUserChannelsParams = z.object({
-		userId: z.number({ coerce: true }).int().gt(0),
-	});
+const controller = new Hono<{ Variables: ContextVariables}>();
 
-	fastify.withTypeProvider<ZodTypeProvider>().route({
-		method: "GET",
-		url: baseUserChannelsUrl,
-		schema: {
-			params: baseUserChannelsParams,
-			querystring: paginationSchema,
-			response: {
-				200: resultSuccessSchema(counted(z.array(ChannelModel.channelSchema))),
-				404: resultFailureSchema,
-			}
-		},
-		onRequest: fastify.authorizeJWT,
-		async handler(req, reply) {
-			const { userId } = req.params;
-			const { skip, take } = { ...paginationDefaults, ...req.query };
-			const [data, count] = await userToChannelRelationsRepository.listUserChannels(userId, { skip, take });
-			return reply.send(result({ data, count  }));
-		}
-	});
+controller.get(
+	"/",
+	zValidator("param", userIdParamsSchema, paramErrorHook),
+	zValidator("query", pageinationQuerySchema, validationErrorHook),
+	async (c) => {
+		const userToChannelRelationsRepository = di.inject("UserToChannelRelationsRepository");
+		const { userId } = c.req.valid("param");
+		const { skip, take } = { ...paginationDefaults, ...c.req.valid("query") };
+		const [data, count] = await userToChannelRelationsRepository.listUserChannels(userId, { skip, take });
+		return c.json({ data, count } satisfies Counted<ChannelModel.Channel[]>);
+	}
+);
 
-	fastify.withTypeProvider<ZodTypeProvider>().route({
-		method: "GET",
-		url: '/users/:userId/available-channels',
-		schema: {
-			params: baseUserChannelsParams,
-			querystring: paginationSchema,
-			response: {
-				200: resultSuccessSchema(z.array(ChannelModel.channelSchema)),
-				404: resultFailureSchema,
-			}
-		},
-		onRequest: fastify.authorizeJWT,
-		async handler(req, reply) {
-			const { userId } = req.params;
-			const { skip, take } = { ...paginationDefaults, ...req.query };
-			const data = await userToChannelRelationsRepository.listAvailableUnsubscribedChannelsForUser(userId, { skip, take });
-			return reply.send(result(data));
-		}
-	});
+controller.get(
+	"/available-channels",
+	zValidator("param", userIdParamsSchema, paramErrorHook),
+	zValidator("query", pageinationQuerySchema, validationErrorHook),
+	async (c) => {
+		const userToChannelRelationsRepository = di.inject("UserToChannelRelationsRepository");
+		const { userId } = c.req.valid("param");
+		const { skip, take } = { ...paginationDefaults, ...c.req.valid("query") };
+		const data = await userToChannelRelationsRepository.listAvailableUnsubscribedChannelsForUser(userId, { skip, take });
+		return c.json(data satisfies ChannelModel.Channel[]);
+	}
+);
 
-	fastify.withTypeProvider<ZodTypeProvider>().route({
-		method: "POST",
-		url: baseUserChannelsUrl,
-		schema: {
-			params: baseUserChannelsParams,
-			body: z.object({ id: z.number({ coerce: true }).int().gt(0) }),
-			response: {
-				200: resultSuccessSchema(z.null()),
-				404: resultFailureSchema,
-			},
-		},
-		onRequest: fastify.authorizeJWT,
-		async handler(req, reply) {
-			const { userId } = req.params;
-			const channelId = req.body.id;
-			await userToChannelRelationsRepository.connectUserChannel(userId, channelId);
-			fastify.log.info(`Channel added to user ${req.params.userId} edit by ${req.user.id}-${req.user.name}`, req.body);
-			return reply.send(result(null));
-		}
-	});
+controller.post(
+	"/",
+	zValidator("param", userIdParamsSchema, paramErrorHook),
+	zValidator("json", z.object({ id: z.number({ coerce: true }).int().gt(0) }), validationErrorHook),
+	async (c) => {
+		const logger = di.inject("logger");
+		const userToChannelRelationsRepository = di.inject("UserToChannelRelationsRepository");
+		const { userId } = c.req.valid("param");
+		const { id: channelId } = c.req.valid("json");
+		await userToChannelRelationsRepository.connectUserChannel(userId, channelId);
+		logger.info(`Channel added to user ${userId} edit by ${c.get("user").id}-${c.get("user").id}`, channelId);
+		return c.json(null);
+	}
+);
 
-	fastify.withTypeProvider<ZodTypeProvider>().route({
-		method: "DELETE",
-		url: baseUserChannelsUrl,
-		schema: {
-			querystring: z.object({ id: batchIdsSchema }),
-			params: z.object({
-				userId: z.number({ coerce: true }).int().gt(0),
-			}),
-			response: {
-				200: resultSuccessSchema(batchOperationStatsSchema),
-				404: resultFailureSchema,
-			},
-		},
-		onRequest: fastify.authorizeJWT,
-		async handler(req, reply) {
-			const { userId } = req.params;
-			const ids = parseIds(req.query.id);
+controller.delete(
+	"/",
+	zValidator("param", userIdParamsSchema, paramErrorHook),
+	zValidator("query",  z.object({ id: batchIdsSchema }), validationErrorHook),
+	async (c) => {
+		const userToChannelRelationsRepository = di.inject("UserToChannelRelationsRepository");
+		const { userId } = c.req.valid("param");
+		const ids = parseIds(c.req.valid("query").id);
 
-			const count = await userToChannelRelationsRepository.disconnectUserChannels(userId, ids)
-				// .catch(handlerDbNotFound(userNotFound(userId)))
+		const count = await userToChannelRelationsRepository.disconnectUserChannels(userId, ids)
+			// .catch(handlerDbNotFound(userNotFound(userId)))
 
-			const data = {
-				count,
-				outOf: ids.length,
-			};
-			return reply.send(result(data));
-		}
-	});
-}
+		const data = {
+			count,
+			outOf: ids.length,
+		};
+		return c.json(data satisfies BatchOperationStats);
+	}
+)
+
+export default controller;
