@@ -1,77 +1,53 @@
+import { Hono } from 'hono';
 import z from "zod";
-import type { ZodTypeProvider } from "fastify-type-provider-zod";
-import type { FastifyInstance } from "fastify";
+import { zValidator } from '@hono/zod-validator';
+import { paramErrorHook, validationErrorHook } from 'src/middleware/validationErrorHandlers';
+
+import type { ContextVariables } from 'src/ContextVariables';
 import * as GroupModel from "@shared/models/Group";
-import { result, resultFailureSchema, resultSuccessSchema } from "@shared/models/Result";
-import { inject } from "src/injection";
+import { di } from "src/injection";
 
-export function userGroups<Instace extends FastifyInstance>(fastify: Instace) {
-	const userToGroupRelationsRepository = inject("UserToGroupRelationsRepository");
+import { userIdParamsSchema } from './models';
+import { intGt, toInt } from '@shared/helpers/zodHelpers';
 
-	const baseUserGroupsUrl = "/users/:userId/groups";
-	const baseUserGroupsParams = z.object({
-		userId: z.number({ coerce: true }).int().gt(0),
-	});
+const controller = new Hono<{ Variables: ContextVariables }>();
 
-	fastify.withTypeProvider<ZodTypeProvider>().route({
-		method: "POST",
-		url: baseUserGroupsUrl,
-		schema: {
-			params: baseUserGroupsParams,
-			body: z.object({ name: GroupModel.groupNameSchema }),
-			response: {
-				200: resultSuccessSchema(z.null()),
-				404: resultFailureSchema,
-				409: resultFailureSchema,
-			},
-		},
-		onRequest: fastify.authorizeJWT,
-		async handler(req, reply) {
-			const id = req.params.userId;
-			const name = req.body.name;
-			await userToGroupRelationsRepository.connectGroupToUser(id, name);
-			fastify.log.info(`Group added to user ${req.params.userId} edit by ${req.user.id}-${req.user.name}`, req.body);
-			return reply.send(result(null));
-		}
-	});
+controller.post(
+	'/',
+	zValidator("param", userIdParamsSchema, paramErrorHook),
+	zValidator("json", z.object({ name: GroupModel.groupNameSchema }), validationErrorHook),
+	async (c) => {
+		const logger = di.inject("logger");
+		const userToGroupRelationsRepository = di.inject("UserToGroupRelationsRepository");
+		const { userId } = c.req.valid("param");
+		const { name } = c.req.valid("json");
+		await userToGroupRelationsRepository.connectGroupToUser(userId, name);
+		logger.info(`Group added to user ${userId} edit by ${c.get("user").id}-${c.get("user").name}`, name);
+		return c.json(null);
+	}
+);
 
-	fastify.withTypeProvider<ZodTypeProvider>().route({
-		method: "DELETE",
-		url: `${baseUserGroupsUrl}/:groupId`,
-		schema: {
-			params: baseUserGroupsParams.extend({
-				groupId: z.number({ coerce: true }).int().gt(0),
-			}),
-			response: {
-				200: resultSuccessSchema(z.null()),
-				404: resultFailureSchema,
-				409: resultFailureSchema,
-			},
-		},
-		onRequest: fastify.authorizeJWT,
-		async handler(req, reply) {
-			const { userId, groupId } = req.params;
-			await userToGroupRelationsRepository.deleteGroupFromUser(userId, groupId);
-			return reply.send(result(null));
-		}
-	});
+controller.delete(
+	"/:groupId",
+	zValidator("param", userIdParamsSchema.extend({ 
+		groupId: z.string().refine(...intGt(0)).transform(toInt) 
+	}), paramErrorHook),
+	async (c) => {
+		const userToGroupRelationsRepository = di.inject("UserToGroupRelationsRepository");
+		const { userId, groupId } = c.req.valid("param");
+		await userToGroupRelationsRepository.deleteGroupFromUser(userId, groupId);
+		return c.json(null);
+	}
+);
 
-	fastify.withTypeProvider<ZodTypeProvider>().route({
-		method: "DELETE",
-		url: baseUserGroupsUrl,
-		schema: {
-			params: baseUserGroupsParams,
-			response: {
-				200: resultSuccessSchema(z.null()),
-				404: resultFailureSchema,
-				409: resultFailureSchema,
-			},
-		},
-		onRequest: fastify.authorizeJWT,
-		async handler(req, reply) {
-			const { userId } = req.params;
-			await userToGroupRelationsRepository.deleteAllGroupsFromUser(userId);
-			return reply.send(result(null));
-		}
-	});
-}
+controller.delete(
+	"/",
+	zValidator("param", userIdParamsSchema, paramErrorHook),
+	async (c) => {
+		const userToGroupRelationsRepository = di.inject("UserToGroupRelationsRepository");
+		const { userId } = c.req.valid("param");
+		await userToGroupRelationsRepository.deleteAllGroupsFromUser(userId);
+		return c.json(null);
+	}
+)
+export default controller;
