@@ -1,14 +1,14 @@
 export class AppListenService {
 	listen: () => void;
-	private _prms: Promise<void>;
+	#prms: Promise<void>;
 
 	constructor() {
 		const defer = Promise.withResolvers<void>();
 		this.listen = defer.resolve;
-		this._prms = defer.promise;
+		this.#prms = defer.promise;
 	}
 
 	async listening(): Promise<void> {
-		return this._prms;
+		return this.#prms;
 	}
 }
