@@ -1,11 +1,12 @@
 import { join } from "path";
-import { writeFile, readFile, access, constants } from "fs/promises";
+import { writeFile, readFile, access, constants, unlink } from "fs/promises";
 import { stripComments } from "jsonc-parser";
 import { serverConfigSchema, type ServerConfig } from "@shared/models/ServerConfig";
 import { Emitter } from "src/util/Emitter";
 import { Lock } from "src/util/Lock";
 import { watchFile } from 'node:fs';
-import { inject } from "src/injection";
+import { di } from "src/injection";
+
 import { getRootDir } from "src/util/getRootDir";
 import { DatabaseConfigurator } from "src/db/DatabaseConnectionTester";
 import { ResultError } from "@shared/models/Result";
@@ -33,7 +34,7 @@ export class SettingsService {
 		private readonly databaseConfigurator = new DatabaseConfigurator(),
 	) {
 		const watchHandler = () => {
-			const log = inject("logger");
+			const log = di.inject("logger");
 			log.warn("Settings file was changed");
 			this.loadConfig().catch((e) => {
 				log.warn("Unable to load settings file", e);
@@ -66,7 +67,7 @@ export class SettingsService {
 		serverConfigSchema.parse(config);
 		const isDbOk = await this.databaseConfigurator.checkConnectionString(config.databaseUrl);
 		if (!isDbOk) {
-			throw new ResultError(400, "Cannot connect to the DB with the provided 'databaseUrl' string");
+			throw new ResultError(422, "Cannot connect to the DB with the provided 'databaseUrl' string");
 		}
 		try {
 			await this.disposerLock.wait();
@@ -82,6 +83,11 @@ export class SettingsService {
 		return this.databaseConfigurator.checkConnectionString(connectionString);
 	}
 
+	async removeConfig(): Promise<void> {
+		await unlink(this.getConfigFileName());
+		this.config = undefined;
+	}
+
 	subscribe(
 		cb: (config?: ServerConfig, oldConfig?: ServerConfig) => MaybePromise<Disposer | void>,
 		fields?: Array<keyof ServerConfig>
@@ -89,18 +95,15 @@ export class SettingsService {
 		let disposer: Disposer | void;
 		return this.emitter.on("change", async (config, oldConfig) => {
 			if (typeof disposer === "function") {
-				const unlock = this.disposerLock.lock();
-				try {
-					// we're waiting for an old disposer to finish prior to launching the new handler
-					await disposer();
-				} finally {
-					unlock();
-				}
+				using _lock = this.disposerLock.lock();
+				// we're waiting for an old disposer to finish prior to launching the new handler
+				await disposer();
 			}
 			const shouldCall = fields?.some((field) => config?.[field] !== oldConfig?.[field]) ?? true;
 			disposer = shouldCall ? await cb(config, oldConfig) : undefined;
 		});
 	}
+
 	unsubscribeAll(): void {
 		this.emitter.clear();
 	}
